Fix bug where scaler wouldn't scale under certain conditions
These conditions seem to be very specific, but basically if the
aspect ratio was close to square or wider AND we loaded the level from
the next level button it wouldn't scale down because the children that
got added to the word panel weren't laid out yet. The fix was to ensure
the words got laid out every time before we laid out the word panel
itself.